+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Ensure emails have name, destination and subject
 * Fix bug in TP account deletion
+* Adds a 'last_extended' attribute to the Project type
 * UI improvements on projects page
 * Fix database creation validation confirmation email
 * Fix database creation request email title